Resolved echo "$1TEST" testcase
