+++ includes/minishell.h
@@ -6,7 +6,7 @@
 /*   By: axlee <axlee@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/10 13:39:49 by gyong-si          #+#    #+#             */
-/*   Updated: 2024/06/12 13:55:10 by axlee            ###   ########.fr       */
+/*   Updated: 2024/06/12 16:36:29 by axlee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -86,7 +86,7 @@ typedef struct s_shell
 	int				prev_fd;
 	pid_t			process_ids[100];
 	int				process_count;
-	int				last_return ;
+	int last_return ;
 	int				flag;
 	int				signal_received;
 	bool			end;
@@ -144,10 +144,18 @@ void				handle_variable_value(char **result, t_shell *minishell,
 						char *var_value);
 void				handle_variable_not_found(char **result, t_shell *minishell,
 						const char *var_name);
+
+// parser(parse_double_quotes_utils_2)
+void				extract_var_name(char *str, t_shell *minishell,
+						char *var_name, int *var_len);
+void				handle_digit_case(char *str, char **result,
+						t_shell *minishell);
+void				remove_trailing_double_quote(char **result,
+						t_shell *minishell);
 void				handle_env_variable_expansion(char *str, char **result,
 						t_shell *minishell);
 
-// parser(parse_double_quotes_utils_2)
+// parser(parse_double_quotes_utils_3)
 void				handle_special_env_variable(char *result,
 						t_shell *minishell);
 void				extract_variable_name(char *str, char **var_name,
@@ -157,7 +165,7 @@ void				handle_regular_env_variable(char *str, char *result,
 void				handle_env_variable_expansion(char *str, char **result,
 						t_shell *minishell);
 
-// parser (parse_double_quotes_utils_3)
+// parser (parse_double_quotes_utils_4)
 void				process_exit_status(char **result, t_shell *minishell);
 void				process_special_characters(char *str, char **result,
 						t_shell *minishell);
@@ -168,7 +176,7 @@ void				process_special_dollar_cases(char *str, char **result,
 void				process_dollar_character(char *str, char **result,
 						t_shell *minishell);
 
-// parser (parse_double_quotes_utils_4)
+// parser (parse_double_quotes_utils_5)
 void				update_parse_variables(int *len, char **result);
 void				initialize_parse_variables(t_token *token,
 						t_shell *minishell);
