+++ src/builtins/builtin_echo.c
@@ -6,7 +6,7 @@
 /*   By: axlee <axlee@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/12 09:53:07 by axlee             #+#    #+#             */
-/*   Updated: 2024/06/12 15:28:42 by gyong-si         ###   ########.fr       */
+/*   Updated: 2024/06/12 16:06:06 by axlee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
