+++ src/parser/parse_double_quotes.c
@@ -6,7 +6,7 @@
 /*   By: axlee <axlee@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/06/09 11:00:03 by axlee             #+#    #+#             */
-/*   Updated: 2024/06/12 12:18:38 by gyong-si         ###   ########.fr       */
+/*   Updated: 2024/06/12 16:08:32 by axlee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -17,7 +17,7 @@ void	parse_double_quotes(t_token *token, t_shell *minishell)
 	char	*str;
 	char	*result;
 
-	printf("in parse dq\n");
+	printf("in parse double quotes\n");
 	minishell->i = 0;
 	minishell->j = 0;
 	initialize_parse_variables(token, minishell);
