+++ src/parser/parse_double_quotes_utils_2.c
@@ -1,62 +1,60 @@
 /* ************************************************************************** */
 /*                                                                            */
 /*                                                        :::      ::::::::   */
-/*   parse_double_quotes_utils_1.c                      :+:      :+:    :+:   */
-/*                                                    +:+         +:+     */
+/*   parse_double_quotes_utils_2.c                      :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
 /*   By: axlee <axlee@student.42.fr>                +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
-/*   Created: 2024/06/09 11:46:06 by axlee             #+#    #+#             */
-/*   Updated: 2024/06/09 20:49:45 by axlee            ###   ########.fr       */
+/*   Created: 2024/06/12 16:34:27 by axlee             #+#    #+#             */
+/*   Updated: 2024/06/12 16:38:29 by axlee            ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "minishell.h"
 
-void	handle_special_env_variable(char *result, t_shell *minishell)
+void	extract_var_name(char *str, t_shell *minishell, char *var_name,
+		int *var_len)
 {
-	char	*expanded;
-	int		i;
-	int		j;
-
-	i = 0;
-	j = 0;
-	expanded = ft_itoa(minishell->last_return);
-	ft_strcpy(&result[j], expanded);
-	j += ft_strlen(expanded);
-	free(expanded);
-	i += 2;
+	while (ft_isalnum(str[minishell->i]) || str[minishell->i] == '_')
+		var_name[(*var_len)++] = str[minishell->i++];
+	var_name[*var_len] = '\0';
 }
 
-void	extract_variable_name(char *str, char **var_name, int *var_len)
+void	handle_digit_case(char *str, char **result, t_shell *minishell)
 {
-	char	*var_start;
-	char	*var_end;
-
-	var_start = &str[1];
-	var_end = var_start;
-	while (*var_end && (ft_isalnum(*var_end) || *var_end == '_'))
-		var_end++;
-	*var_len = var_end - var_start;
-	*var_name = ft_strndup(var_start, *var_len);
+	minishell->i++;
+	while (str[minishell->i] != '\0' && str[minishell->i] != ' ')
+		(*result)[minishell->j++] = str[minishell->i++];
 }
 
-void	handle_regular_env_variable(char *str, char *result, t_shell *minishell)
+void	remove_trailing_double_quote(char **result, t_shell *minishell)
 {
-	char	*var_name;
-	char	*expanded;
-	int		var_len;
-	int		i;
-	int		j;
-
-	i = 0;
-	j = 0;
-	extract_variable_name(str, &var_name, &var_len);
-	expanded = get_env_value(minishell, var_name);
-	if (expanded)
+	if ((*result)[minishell->j - 1] == '"')
 	{
-		ft_strcpy(&result[j], expanded);
-		j += strlen(expanded);
+		(*result)[minishell->j - 1] = '\0';
+		minishell->j--;
 	}
-	i += var_len;
-	free(var_name);
+}
+
+void	handle_env_variable_expansion(char *str, char **result,
+		t_shell *minishell)
+{
+	char	var_name[256];
+	char	*var_value;
+	int		var_len;
+
+	var_len = 0;
+	minishell->i++;
+	if (!isdigit(str[minishell->i]))
+	{
+		extract_var_name(str, minishell, var_name, &var_len);
+		var_value = get_env_value(minishell, var_name);
+		if (var_value)
+			handle_variable_value(result, minishell, var_value);
+		else
+			handle_variable_not_found(result, minishell, var_name);
+	}
+	else
+		handle_digit_case(str, result, minishell);
+	remove_trailing_double_quote(result, minishell);
 }
